Optimize drawing a little bit by caching TCODColor.black
We use the color black quite a lot, once per onscreen tile, so we can
get a bit of a speed boost by not crossing the P/Invoke border so
many times per drawn frame and instead only cross over for non-black
colors.

It may be worth creating a CLR-side map of cached colors from TCOD,
but this may not be worth the effort just now.