Initial work on core of userinterface
Add OctoGhast.UserInterface to SLN
Initial draft of Canvas and Widget classes.

Canvas provides a drawing surface that supports
extra options for colouring and some functions
not present in TCODConsole.

Widget provides a base class to derive controls
from.

A pigment wraps together a foreground, background, and background flag. Stored in a dictionary with
keys for various parts of a control, we can load them in from JSON or something.

Templates provide a way to influence a widgets drawing behaviour,
more framework for themes really.

Add an implicit cast from Color to TCODColor