Switch to using a Pigment type for color selections.
Pigments bind together a foreground, background, and background flag; Allowing
easier passing around of color information.

PigmentMapping provides a way to have a set of pigments with an optional set of
overrides for parts of the UI.

Each UI part is given an entry in the PigmentType enumeration which is used as a
key into PigmentMapping. Controls can use these to draw appropriately.

Also a reformat on Canvas, which slightly obscures the changes for pigments.
PigmentSession is introduced, which provides a simple way to introduce a colour
scope, any canvas operations done from within a PigmentSession will use the passed
pigments, then reverse back to the defaults.