Lots of changes and a working UI framework ready to expand.
Further commits due to clean up the stuff not being used, in another feature branch.

This adds the basis of the framework for UI controls with the only working control right now,
a magical Button.

The API will be cleaned up soon, some of the interaction between the core and the ui assembly are
a bit hacky and need to be replaced with something cleaner. Like Config needing manual setup.

This also introduces an Application framework, which makes it simple to dispatch events around to
make everything work.