Add basic framework for BindingTargets
Allows using INotifyPropertyChanged to setup bindings for controls.
Controls can use this to be notified when a dependent property changes
so the UI updates automatically. (WPF Style Bindings, essentially)

Adds support to the ControlBase, so inherited controls can use
SetupBinding<T>(this, () => Property) to indicate that 'Property' on
the control is the target site for the binding given to this control.

The initial binding information is passed to the ControlTemplate
used to construct the actual control.