Fix a bug in the Canvas Blit()
It erroneously provided an alpha value to BlitToConsole which calls
BlitAlpha on the RLConsole instance, this causes the feature that
controls clear their background when drawing to fail. Because space
characters are treated specially when drawing in alpha mode.

This provides a non-alpha version for when the user hasn't provided an
alpha value, fixing most of the controls drawing (Panels, Buttons, etc)
behaviour with regards to showing controls 'underneath' them