Refactor LoadGameScreen
Rearrange the positioning of controls, now that controls aren't
always alpha-blended to the background, the labels need to be
a tiny bit shorter to not overdraw the frame from the panel.

Add stub code for loading/deleting worlds.

Utilize a new QueryYN function under UIHelper that simulates pausing
the game and waiting for a user response to a query.